requirements.txt
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
breadability==0.1.20
cachetools==6.1.0
catalogue==2.0.10
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
docopt==0.6.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
llvmlite==0.44.0
lxml==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
narwhals==1.45.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.0
packaging==25.0
pandas==2.3.0
patsy==1.0.1
piexif==1.1.3
pillow==11.3.0
plotly==6.2.0
plotly-express==0.4.1
preshed==3.0.10
protobuf==6.31.1
pyarrow==20.0.0
pycountry==24.6.1
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
pypdf==5.7.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rake-nltk==1.0.6
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==80.9.0
shap==0.48.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smart_open==7.3.0.post1
smmap==5.0.2
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
statsmodels==0.14.4
streamlit==1.46.1
sumy==0.11.0
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.53.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.2
xgboost==3.0.2
